stk(3): fix argument type of exception callback function
The callback function for malloc(3) failure in the stk(3) routines
(which can be specified using stkinstall() and is set to nomemory()
in init.c) uses a size argument of type int. That's a type mismatch
with all the other size arguments and variables in stk(3) which use
size_t, an unsigned type that may be larger than int.

This is all quite inconsequential as nothing in our code base (or
in the complete old AT&T AST code base) actually uses that size for
anything, but it's still wrong, so this corrects the interface.

With this very minor API change, let's bump the libast API version
to 20220801, the date of the upcoming ksh 93u+m 1.0.0 release. :)

The ksh93/sh/init.c nomemory() function now reports the size that
could not be allocated, just because it can.

+++ src/lib/libast/include/stk.h
@@ -45,7 +45,7 @@
 extern Sfio_t		_Stk_data;
 
 extern Stk_t*		stkopen(int);
-extern Stk_t*		stkinstall(Stk_t*, char*(*)(int));
+extern Stk_t*		stkinstall(Stk_t*, char*(*)(size_t));
 extern int		stkclose(Stk_t*);
 extern unsigned int	stklink(Stk_t*);
 extern char*		stkalloc(Stk_t*, size_t);
